+++ phenol-analysis/src/test/java/org/monarchinitiative/phenol/stats/BonferroniTest.java
@@ -2,7 +2,6 @@ package org.monarchinitiative.phenol.stats;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.monarchinitiative.phenol.ontology.data.TermId;
 
 
 import java.util.List;